requirements.txt
anthropic==0.91.0
python-dotenv==1.2.1
streamlit==1.50.0
